Simplify ZSA implementation (#178)
- Update comments
- Reduce diff with upstream
- Unify Rust version to 1.71 (Cargo.toml, README.md and
rust-toolchain.toml)
- Remove pinned version for half and font-kit in Cargo.toml
- Remove public unsplittable function
- Clean tests/zsa.rs by creating one Merkle tree with all notes and by
generating new keys to avoid using the same address for both sender and
recipient in bundles
- Simplify `IssueAction::verify(...)` by moving the
`AssetBaseCannotBeIdentityPoint` check outside the loop.
- Rationale: Previously we checked that for each note, its asset is not
equal to identity point and for each note, its asset is equal to the
issue asset. Now we checked that the issue asset is not equal to the
identity point and for each note, its asset is equal to the issue asset.
Those two versions will fail in the same cases.
- We removed `action_verify_invalid_for_asset_base_as_identity`,
`issue_bundle_cannot_be_signed_with_asset_base_identity_point` and
`issue_bundle_verify_fail_asset_base_identity_point` tests because it is
now very difficult to create a test for this case. To create a test for
this case, we have to choose an `ik` and an `asset_desc_hash` such that
the `AssetBase` derived from them is equal to the identity point.